Moves files, adds some docs
