+++ apps/frontend/src/app/api/model/model.type.ts
@@ -1,3 +1,8 @@
 import { Dto } from './dto.base';
 
+/**
+ * Model type for Dtos. Removes the `id` property.
+ *
+ * @template T - The type of the DTO
+ */
 export type Model<T extends Dto> = Omit<T, Dto['id']>;
