+++ apps/frontend/src/app/app.routes.ts
@@ -1,9 +1,9 @@
 import { Route } from '@angular/router';
 import { isAuthenticatedOrRedirectToLogin } from './auth/auth.guard';
-import { DashboardComponent } from './components/organisms/dashboard.component';
-import { LoginComponent } from './components/organisms/login.component';
-import { UsersComponent } from './components/organisms/users.component';
+import { DashboardComponent } from './components/pages/dashboard.component';
+import { LoginComponent } from './components/pages/login.component';
 import { SampleDesignComponent } from './components/pages/sample-design.component';
+import { UsersComponent } from './components/pages/users.component';
 
 export const appRoutes: Route[] = [
   {
