+++ apps/frontend/src/app/components/molecules/user.component.ts
