+++ apps/frontend/src/app/components/pages/dashboard.component.ts
