+++ apps/frontend/src/app/components/templates/app.component.ts
