+++ apps/frontend/src/main.ts
@@ -1,7 +1,8 @@
 import { bootstrapApplication } from '@angular/platform-browser';
 import { appConfig } from './app/app.config';
-import { AppComponent } from './app/components/pages/app.component';
+import { AppComponent } from './app/components/templates/app.component';
 
 bootstrapApplication(AppComponent, appConfig).catch((err) =>
   console.error(err)
 );
+
