+++ apps/frontend/src/styles.scss
@@ -4,8 +4,6 @@
   @include nb-theme-global();
 }
 
-@import url(./styles/vars.scss);
-
 .__emtpy__ {
   display: none;
 }
