--- apps/frontend/src/styles/vars.scss
@@ -1,50 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-$color-primary-lighter-4: #F2FEF9;
-$color-primary-lighter-3: #E5FDF5;
-$color-primary-lighter-2: #D6FAF1;
-$color-primary-lighter-1: #CAF6EE;
-$color-primary: #B7F0EC;
-$color-primary-darker-1: #85CCCE;
-$color-primary-darker-2: #5CA3AC;
-$color-primary-darker-3: #3A7B8B;
-$color-primary-darker-4: #235B73;
-
-$color-success-lighter-4: #CBFCDD;
-$color-success-lighter-3: #99FAC5;
-$color-success-lighter-2: #65F2B2;
-$color-success-lighter-1: #3EE6AA;
-$color-success: #06D6A0;
-$color-success-darker-1: #04B89A;
-$color-success-darker-2: #039A90;
-$color-success-darker-3: #01787C;
-$color-success-darker-4: #015966;
-
-$color-info-lighter-4: #CFFCFB;
-$color-info-lighter-3: #A0F7FA;
-$color-info-lighter-2: #70E4F2;
-$color-info-lighter-1: #4BCAE6;
-$color-info: #16A6D6;
-$color-info-darker-1: #1082B8;
-$color-info-darker-2: #0B629A;
-$color-info-darker-3: #07467C;
-$color-info-darker-4: #043266;
-
-$color-warning-lighter-4: #FFF9E0;
-$color-warning-lighter-3: #FFF2C1;
-$color-warning-lighter-2: #FFE9A3;
-$color-warning-lighter-1: #FFE08C;
-$color-warning: #FFD166;
-$color-warning-darker-1: #DBAB4A;
-$color-warning-darker-2: #B78733;
-$color-warning-darker-3: #936620;
-$color-warning-darker-4: #7A4E13;
-
-$color-danger-lighter-4: #FEE4DA;
-$color-danger-lighter-3: #FDC4B6;
-$color-danger-lighter-2: #FA9C90;
-$color-danger-lighter-1: #F57674;
-$color-danger: #EF4753;
-$color-danger-darker-1: #CD334C;
-$color-danger-darker-2: #AC2346;
-$color-danger-darker-3: #8A163E;
-$color-danger-darker-4: #720D39;
